requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Vector Store & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Utilities
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3

# Search
elasticsearch==8.12.0

# Graph utilities
networkx==3.2.1

# HTTP Client
httpx==0.26.0
